+++ frontend/src/utils/nested-selection-difference.ts
